jobs/requirements.txt
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyGithub==2.1.1
openai==1.10.0
httpx==0.26.0
tenacity==8.2.3
pydantic==2.5.3
email-validator==2.1.0
schedule==1.2.0
